requirements.txt
google-api-python-client 
isodate
google-generativeai 
python-dotenv
SpeechRecognition
pyaudio
streamlit